requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bidict==0.23.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
distro==1.9.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
numpy==2.2.0
openai==1.57.3
pandas==2.2.3
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
wsproto==1.2.0
xgboost==2.1.3
